resolve withdraw page
